Refactor into separate pokemon api files
