+++ decoder/pokemonRtree.go
@@ -7,138 +7,12 @@ import (
 	"github.com/puzpuzpuz/xsync/v2"
 	log "github.com/sirupsen/logrus"
 	"github.com/tidwall/rtree"
-	"golang.org/x/exp/slices"
 	"golbat/config"
-	"golbat/geo"
 	"gopkg.in/guregu/null.v4"
 	"math"
 	"strconv"
-	"strings"
 	"sync"
-	"time"
 )
-
-type ApiPokemonScan struct {
-	Min             geo.Location                `json:"min"`
-	Max             geo.Location                `json:"max"`
-	Center          geo.Location                `json:"center"`
-	Limit           int                         `json:"limit"`
-	GlobalFilter    *ApiPokemonFilter           `json:"global"`
-	SpecificFilters map[string]ApiPokemonFilter `json:"filters"`
-}
-
-type ApiPokemonScan2 struct {
-	Min        geo.Location           `json:"min"`
-	Max        geo.Location           `json:"max"`
-	Limit      int                    `json:"limit"`
-	DnfFilters []*ApiPokemonDnfFilter `json:"filters"`
-}
-
-type ApiPokemonSearch struct {
-	Min       geo.Location `json:"min"`
-	Max       geo.Location `json:"max"`
-	Center    geo.Location `json:"center"`
-	Limit     int          `json:"limit"`
-	SearchIds []int16      `json:"searchIds"`
-}
-
-type ApiPokemonFilter struct {
-	Iv         []int8                      `json:"iv"`
-	AtkIv      []int8                      `json:"atk_iv"`
-	DefIv      []int8                      `json:"def_iv"`
-	StaIv      []int8                      `json:"sta_iv"`
-	Level      []int8                      `json:"level"`
-	Cp         []int16                     `json:"cp"`
-	Gender     int8                        `json:"gender"`
-	Additional *ApiPokemonAdditionalFilter `json:"additional"`
-	Pvp        *ApiPvpFilter               `json:"pvp"`
-}
-
-type ApiPvpFilter struct {
-	Little []int16 `json:"little"`
-	Great  []int16 `json:"great"`
-	Ultra  []int16 `json:"ultra"`
-}
-
-type ApiPokemonAdditionalFilter struct {
-	IncludeEverything bool `json:"include_everything"`
-	IncludeHundos     bool `json:"include_hundoiv"`
-	IncludeNundos     bool `json:"include_zeroiv"`
-	IncludeXxs        bool `json:"include_xxs"`
-	IncludeXxl        bool `json:"include_xxl"`
-}
-
-type ApiPokemonDnfFilter struct {
-	Pokemon []ApiPokemonDnfId     `json:"pokemon"`
-	Iv      *ApiPokemonDnfMinMax8 `json:"iv"`
-	AtkIv   *ApiPokemonDnfMinMax8 `json:"atk_iv"`
-	DefIv   *ApiPokemonDnfMinMax8 `json:"def_iv"`
-	StaIv   *ApiPokemonDnfMinMax8 `json:"sta_iv"`
-	Level   *ApiPokemonDnfMinMax8 `json:"level"`
-	Cp      *ApiPokemonDnfMinMax  `json:"cp"`
-	Gender  int8                  `json:"gender"`
-	Size    int8                  `json:"size"`
-	Little  *ApiPokemonDnfMinMax  `json:"pvp_little"`
-	Great   *ApiPokemonDnfMinMax  `json:"pvp_great"`
-	Ultra   *ApiPokemonDnfMinMax  `json:"pvp_ultra"`
-}
-
-type ApiPokemonDnfId struct {
-	Pokemon int16  `json:"id"`
-	Form    *int16 `json:"form"`
-}
-
-type ApiPokemonDnfMinMax struct {
-	Min int16 `json:"min"`
-	Max int16 `json:"max"`
-}
-
-type ApiPokemonDnfMinMax8 struct {
-	Min int8 `json:"min"`
-	Max int8 `json:"max"`
-}
-
-type ApiPokemonResult struct {
-	Id                      string                           `json:"id"`
-	PokestopId              null.String                      `json:"pokestop_id"`
-	SpawnId                 null.Int                         `json:"spawn_id"`
-	Lat                     float64                          `json:"lat"`
-	Lon                     float64                          `json:"lon"`
-	Weight                  null.Float                       `json:"weight"`
-	Size                    null.Int                         `json:"size"`
-	Height                  null.Float                       `json:"height"`
-	ExpireTimestamp         null.Int                         `json:"expire_timestamp"`
-	Updated                 null.Int                         `json:"updated"`
-	PokemonId               int16                            `json:"pokemon_id"`
-	Move1                   null.Int                         `json:"move_1"`
-	Move2                   null.Int                         `json:"move_2"`
-	Gender                  null.Int                         `json:"gender"`
-	Cp                      null.Int                         `json:"cp"`
-	AtkIv                   null.Int                         `json:"atk_iv"`
-	DefIv                   null.Int                         `json:"def_iv"`
-	StaIv                   null.Int                         `json:"sta_iv"`
-	Iv                      null.Float                       `json:"iv"`
-	Form                    null.Int                         `json:"form"`
-	Level                   null.Int                         `json:"level"`
-	EncounterWeather        uint8                            `json:"encounter_weather"`
-	Weather                 null.Int                         `json:"weather"`
-	Costume                 null.Int                         `json:"costume"`
-	FirstSeenTimestamp      int64                            `json:"first_seen_timestamp"`
-	Changed                 int64                            `json:"changed"`
-	CellId                  null.Int                         `json:"cell_id"`
-	ExpireTimestampVerified bool                             `json:"expire_timestamp_verified"`
-	DisplayPokemonId        null.Int                         `json:"display_pokemon_id"`
-	IsDitto                 bool                             `json:"is_ditto"`
-	SeenType                null.String                      `json:"seen_type"`
-	Shiny                   null.Bool                        `json:"shiny"`
-	Username                null.String                      `json:"username"`
-	Capture1                null.Float                       `json:"capture_1"`
-	Capture2                null.Float                       `json:"capture_2"`
-	Capture3                null.Float                       `json:"capture_3"`
-	Pvp                     map[string][]gohbem.PokemonEntry `json:"pvp"`
-	IsEvent                 int8                             `json:"is_event"`
-	Distance                float64                          `json:"distance,omitempty"`
-}
 
 type PokemonLookupCacheItem struct {
 	PokemonLookup    *PokemonLookup
@@ -165,12 +39,6 @@ type PokemonPvpLookup struct {
 	Little int16
 	Great  int16
 	Ultra  int16
-}
-
-type Available struct {
-	PokemonId int16 `json:"id"`
-	Form      int16 `json:"form"`
-	Count     int   `json:"count"`
 }
 
 var pokemonLookupCache *xsync.MapOf[uint64, PokemonLookupCacheItem]
@@ -292,552 +160,4 @@ func removePokemonFromTree(pokemon *Pokemon) {
 	if beforeLen != afterLen+1 {
 		log.Infof("PokemonRtree - UNEXPECTED removing %d, lat %f lon %f size %d->%d Map Len %d", pokemonId, pokemon.Lat, pokemon.Lon, beforeLen, afterLen, pokemonLookupCache.Size())
 	}
-}
-
-func GetPokemonInArea(retrieveParameters ApiPokemonScan) []*ApiPokemonResult {
-	// Validate filters
-
-	validateFilter := func(filter *ApiPokemonFilter) bool {
-
-		if filter.StaIv != nil && len(filter.StaIv) != 2 {
-			return false
-		}
-		if filter.AtkIv != nil && len(filter.AtkIv) != 2 {
-			return false
-		}
-		if filter.DefIv != nil && len(filter.DefIv) != 2 {
-			return false
-		}
-		if filter.Iv != nil && len(filter.Iv) != 2 {
-			return false
-		}
-		if filter.Level != nil && len(filter.Level) != 2 {
-			return false
-		}
-		if filter.Cp != nil && len(filter.Cp) != 2 {
-			return false
-		}
-
-		if filter.Pvp != nil {
-			if filter.Pvp.Little != nil && len(filter.Pvp.Little) != 2 {
-				return false
-			}
-			if filter.Pvp.Great != nil && len(filter.Pvp.Great) != 2 {
-				return false
-			}
-			if filter.Pvp.Ultra != nil && len(filter.Pvp.Ultra) != 2 {
-				return false
-			}
-		}
-		return true
-	}
-
-	if retrieveParameters.GlobalFilter != nil && !validateFilter(retrieveParameters.GlobalFilter) {
-		log.Errorf("GetPokemonInArea - Invalid global filter")
-		return nil
-	}
-
-	for _, filter := range retrieveParameters.SpecificFilters {
-		if !validateFilter(&filter) {
-			log.Errorf("GetPokemonInArea - Invalid specific filter")
-			return nil
-		}
-	}
-
-	start := time.Now()
-
-	min := retrieveParameters.Min
-	max := retrieveParameters.Max
-	specificPokemonFilters := retrieveParameters.SpecificFilters
-	globalFilter := retrieveParameters.GlobalFilter
-
-	maxPokemon := config.Config.Tuning.MaxPokemonResults
-	if retrieveParameters.Limit > 0 && retrieveParameters.Limit < maxPokemon {
-		maxPokemon = retrieveParameters.Limit
-	}
-
-	pokemonExamined := 0
-	pokemonSkipped := 0
-
-	isPokemonMatch := func(pokemonLookup *PokemonLookup, pvpLookup *PokemonPvpLookup, filter ApiPokemonFilter) bool {
-		// start with filter true if we have any filter set (no filters no match)
-		filterMatched := filter.Iv != nil || filter.StaIv != nil || filter.AtkIv != nil || filter.DefIv != nil || filter.Level != nil || filter.Cp != nil || filter.Gender != 0
-		pvpMatched := false // assume pvp match is true unless any filter matches
-		additionalMatch := false
-
-		if filterMatched {
-			if filter.Iv != nil && (pokemonLookup.Iv < filter.Iv[0] || pokemonLookup.Iv > filter.Iv[1]) {
-				filterMatched = false
-			} else if filter.StaIv != nil && (pokemonLookup.Sta < filter.StaIv[0] || pokemonLookup.Sta > filter.StaIv[1]) {
-				filterMatched = false
-			} else if filter.AtkIv != nil && (pokemonLookup.Atk < filter.AtkIv[0] || pokemonLookup.Atk > filter.AtkIv[1]) {
-				filterMatched = false
-			} else if filter.DefIv != nil && (pokemonLookup.Def < filter.DefIv[0] || pokemonLookup.Def > filter.DefIv[1]) {
-				filterMatched = false
-			} else if filter.Level != nil && (pokemonLookup.Level < filter.Level[0] || pokemonLookup.Level > filter.Level[1]) {
-				filterMatched = false
-			} else if filter.Cp != nil && (pokemonLookup.Cp < filter.Cp[0] || pokemonLookup.Cp > filter.Cp[1]) {
-				filterMatched = false
-			} else if filter.Gender != 0 && pokemonLookup.Gender != filter.Gender {
-				filterMatched = false
-			}
-		}
-
-		if filter.Additional != nil {
-			if filter.Additional.IncludeEverything {
-				additionalMatch = true
-			} else if filter.Additional.IncludeNundos && pokemonLookup.Sta == 0 && pokemonLookup.Atk == 0 && pokemonLookup.Def == 0 {
-				additionalMatch = true
-			} else if filter.Additional.IncludeHundos && pokemonLookup.Sta == 15 && pokemonLookup.Atk == 15 && pokemonLookup.Def == 15 {
-				additionalMatch = true
-			} else if filter.Additional.IncludeXxs && pokemonLookup.Size == 1 {
-				additionalMatch = true
-			} else if filter.Additional.IncludeXxl && pokemonLookup.Size == 5 {
-				additionalMatch = true
-			}
-		}
-
-		pvpFilter := filter.Pvp
-		if pvpFilter != nil && pvpLookup != nil {
-			if pvpFilter.Little != nil && (pvpLookup.Little >= pvpFilter.Little[0] && pvpLookup.Little <= pvpFilter.Little[1]) {
-				pvpMatched = true
-			}
-			if pvpFilter.Great != nil && (pvpLookup.Great >= pvpFilter.Great[0] && pvpLookup.Great <= pvpFilter.Great[1]) {
-				pvpMatched = true
-			}
-			if pvpFilter.Ultra != nil && (pvpLookup.Ultra >= pvpFilter.Ultra[0] && pvpLookup.Ultra <= pvpFilter.Ultra[1]) {
-				pvpMatched = true
-			}
-		}
-
-		return filterMatched || pvpMatched || additionalMatch
-	}
-
-	pokemonTreeMutex.RLock()
-	pokemonTree2 := pokemonTree.Copy()
-	pokemonTreeMutex.RUnlock()
-
-	lockedTime := time.Since(start)
-
-	performScan := func() (returnKeys []uint64) {
-		pokemonMatched := 0
-		pokemonTree2.Search([2]float64{min.Longitude, min.Latitude}, [2]float64{max.Longitude, max.Latitude},
-			func(min, max [2]float64, pokemonId uint64) bool {
-				pokemonExamined++
-
-				pokemonLookupItem, found := pokemonLookupCache.Load(pokemonId)
-				if !found {
-					pokemonSkipped++
-					// Did not find cached result, something amiss?
-					return true
-				}
-
-				pokemonLookup := pokemonLookupItem.PokemonLookup
-				pvpLookup := pokemonLookupItem.PokemonPvpLookup
-
-				globalFilterMatched := false
-				if globalFilter != nil {
-					globalFilterMatched = isPokemonMatch(pokemonLookup, pvpLookup, *globalFilter)
-				}
-				specificFilterMatched := false
-
-				if !globalFilterMatched && specificPokemonFilters != nil {
-					var formString strings.Builder
-					formString.WriteString(strconv.Itoa(int(pokemonLookup.PokemonId)))
-					formString.WriteByte('-')
-					formString.WriteString(strconv.Itoa(int(pokemonLookup.Form)))
-					filter, found := specificPokemonFilters[formString.String()]
-
-					if found {
-						specificFilterMatched = isPokemonMatch(pokemonLookup, pvpLookup, filter)
-					}
-				}
-
-				if globalFilterMatched || specificFilterMatched {
-					returnKeys = append(returnKeys, pokemonId)
-					pokemonMatched++
-					if pokemonMatched > maxPokemon {
-						log.Infof("GetPokemonInArea - result would exceed maximum size (%d), stopping scan", maxPokemon)
-						return false
-					}
-				}
-
-				return true // always continue
-			})
-
-		return
-	}
-
-	returnKeys := performScan()
-
-	results := make([]*ApiPokemonResult, 0, len(returnKeys))
-
-	for _, key := range returnKeys {
-		if pokemonCacheEntry := pokemonCache.Get(strconv.FormatUint(key, 10)); pokemonCacheEntry != nil {
-			pokemon := pokemonCacheEntry.Value()
-
-			apiPokemon := ApiPokemonResult{
-				Id:              pokemon.Id,
-				PokestopId:      pokemon.PokestopId,
-				SpawnId:         pokemon.SpawnId,
-				Lat:             pokemon.Lat,
-				Lon:             pokemon.Lon,
-				Weight:          pokemon.Weight,
-				Size:            pokemon.Size,
-				Height:          pokemon.Height,
-				ExpireTimestamp: pokemon.ExpireTimestamp,
-				Updated:         pokemon.Updated,
-				PokemonId:       pokemon.PokemonId,
-				Move1:           pokemon.Move1,
-				Move2:           pokemon.Move2,
-				Gender:          pokemon.Gender,
-				Cp:              pokemon.Cp,
-				AtkIv:           pokemon.AtkIv,
-				DefIv:           pokemon.DefIv,
-				StaIv:           pokemon.StaIv,
-				//not IvInactive
-				Iv:                      pokemon.Iv,
-				Form:                    pokemon.Form,
-				Level:                   pokemon.Level,
-				EncounterWeather:        pokemon.EncounterWeather, //? perhaps do not include
-				Weather:                 pokemon.Weather,
-				Costume:                 pokemon.Costume,
-				FirstSeenTimestamp:      pokemon.FirstSeenTimestamp,
-				Changed:                 pokemon.Changed,
-				CellId:                  pokemon.CellId,
-				ExpireTimestampVerified: pokemon.ExpireTimestampVerified,
-				DisplayPokemonId:        pokemon.DisplayPokemonId,
-				IsDitto:                 pokemon.IsDitto,
-				SeenType:                pokemon.SeenType,
-				Shiny:                   pokemon.Shiny,
-				Username:                pokemon.Username,
-				Pvp: func() map[string][]gohbem.PokemonEntry {
-					if ohbem != nil {
-						pvp, err := ohbem.QueryPvPRank(int(pokemon.PokemonId),
-							int(pokemon.Form.ValueOrZero()),
-							int(pokemon.Costume.ValueOrZero()),
-							int(pokemon.Gender.ValueOrZero()),
-							int(pokemon.AtkIv.ValueOrZero()),
-							int(pokemon.DefIv.ValueOrZero()),
-							int(pokemon.StaIv.ValueOrZero()),
-							float64(pokemon.Level.ValueOrZero()))
-						if err != nil {
-							return nil
-						}
-						return pvp
-					}
-					return nil
-				}(),
-			}
-
-			results = append(results, &apiPokemon)
-		}
-	}
-
-	log.Infof("GetPokemonInArea - total time %s (locked time %s), %d scanned, %d skipped, %d returned", time.Since(start), lockedTime, pokemonExamined, pokemonSkipped, len(results))
-
-	return results
-}
-
-func GetPokemonInArea2(retrieveParameters ApiPokemonScan2) []*ApiPokemonResult {
-	type dnfFilterLookup struct {
-		pokemon int16
-		form    int16
-	}
-
-	dnfFilters := make(map[dnfFilterLookup][]*ApiPokemonDnfFilter)
-
-	for _, filter := range retrieveParameters.DnfFilters {
-		if len(filter.Pokemon) > 0 {
-			for _, keyString := range filter.Pokemon {
-				var pokemonId int16 = -1
-				if keyString.Pokemon == 0 {
-					pokemonId = -1
-				}
-				var formId int16 = -1
-				if keyString.Form != nil {
-					formId = *keyString.Form
-				}
-				key := dnfFilterLookup{
-					pokemon: pokemonId,
-					form:    formId,
-				}
-				dnfFilters[key] = append(dnfFilters[key], filter)
-			}
-		} else {
-			key := dnfFilterLookup{
-				pokemon: -1,
-				form:    -1,
-			}
-			dnfFilters[key] = append(dnfFilters[key], filter)
-		}
-	}
-
-	start := time.Now()
-
-	minLocation := retrieveParameters.Min
-	maxLocation := retrieveParameters.Max
-
-	maxPokemon := config.Config.Tuning.MaxPokemonResults
-	if retrieveParameters.Limit > 0 && retrieveParameters.Limit < maxPokemon {
-		maxPokemon = retrieveParameters.Limit
-	}
-
-	pokemonExamined := 0
-	pokemonSkipped := 0
-
-	isPokemonDnfMatch := func(pokemonLookup *PokemonLookup, pvpLookup *PokemonPvpLookup, filter *ApiPokemonDnfFilter) bool {
-		if filter.Iv != nil && (pokemonLookup.Iv < filter.Iv.Min || pokemonLookup.Iv > filter.Iv.Max) ||
-			filter.StaIv != nil && (pokemonLookup.Sta < filter.StaIv.Min || pokemonLookup.Sta > filter.StaIv.Max) ||
-			filter.AtkIv != nil && (pokemonLookup.Atk < filter.AtkIv.Min || pokemonLookup.Atk > filter.AtkIv.Max) ||
-			filter.DefIv != nil && (pokemonLookup.Def < filter.DefIv.Min || pokemonLookup.Def > filter.DefIv.Max) ||
-			filter.Level != nil && (pokemonLookup.Level < filter.Level.Min || pokemonLookup.Level > filter.Level.Max) ||
-			filter.Cp != nil && (pokemonLookup.Cp < filter.Cp.Min || pokemonLookup.Cp > filter.Cp.Max) ||
-			filter.Gender != 0 && pokemonLookup.Gender != filter.Gender ||
-			filter.Size != 0 && pokemonLookup.Size != filter.Size {
-			return false
-		}
-
-		if pvpLookup == nil ||
-			filter.Little != nil && (pvpLookup.Little < filter.Little.Min || pvpLookup.Little > filter.Little.Max) ||
-			filter.Great != nil && (pvpLookup.Great < filter.Great.Min || pvpLookup.Great > filter.Great.Max) ||
-			filter.Ultra != nil && (pvpLookup.Ultra < filter.Ultra.Min || pvpLookup.Ultra > filter.Ultra.Max) {
-			return false
-		}
-		return true
-	}
-
-	pokemonTreeMutex.RLock()
-	pokemonTree2 := pokemonTree.Copy()
-	pokemonTreeMutex.RUnlock()
-
-	lockedTime := time.Since(start)
-
-	performScan := func() (returnKeys []uint64) {
-		pokemonMatched := 0
-		pokemonTree2.Search([2]float64{minLocation.Longitude, minLocation.Latitude}, [2]float64{maxLocation.Longitude, maxLocation.Latitude},
-			func(min, max [2]float64, pokemonId uint64) bool {
-				pokemonExamined++
-
-				pokemonLookupItem, found := pokemonLookupCache.Load(pokemonId)
-				if !found {
-					pokemonSkipped++
-					// Did not find cached result, something amiss?
-					return true
-				}
-
-				pokemonLookup := pokemonLookupItem.PokemonLookup
-				pvpLookup := pokemonLookupItem.PokemonPvpLookup
-
-				matched := false
-
-				filters, found := dnfFilters[dnfFilterLookup{
-					pokemon: pokemonLookup.PokemonId,
-					form:    pokemonLookup.Form}]
-
-				if !found {
-					filters, found = dnfFilters[dnfFilterLookup{
-						pokemon: pokemonLookup.PokemonId,
-						form:    -1}]
-
-					if !found {
-						filters, found = dnfFilters[dnfFilterLookup{
-							pokemon: -1,
-							form:    -1}]
-
-						if !found {
-							return true
-						}
-					}
-				}
-
-				for x := 0; x < len(filters); x++ {
-					if isPokemonDnfMatch(pokemonLookup, pvpLookup, filters[x]) {
-						matched = true
-						break
-					}
-				}
-
-				if matched {
-					returnKeys = append(returnKeys, pokemonId)
-					pokemonMatched++
-					if pokemonMatched > maxPokemon {
-						log.Infof("GetPokemonInArea - result would exceed maximum size (%d), stopping scan", maxPokemon)
-						return false
-					}
-				}
-
-				return true // always continue
-			})
-
-		return
-	}
-
-	returnKeys := performScan()
-
-	results := make([]*ApiPokemonResult, 0, len(returnKeys))
-
-	for _, key := range returnKeys {
-		if pokemonCacheEntry := pokemonCache.Get(strconv.FormatUint(key, 10)); pokemonCacheEntry != nil {
-			pokemon := pokemonCacheEntry.Value()
-
-			apiPokemon := ApiPokemonResult{
-				Id:              pokemon.Id,
-				PokestopId:      pokemon.PokestopId,
-				SpawnId:         pokemon.SpawnId,
-				Lat:             pokemon.Lat,
-				Lon:             pokemon.Lon,
-				Weight:          pokemon.Weight,
-				Size:            pokemon.Size,
-				Height:          pokemon.Height,
-				ExpireTimestamp: pokemon.ExpireTimestamp,
-				Updated:         pokemon.Updated,
-				PokemonId:       pokemon.PokemonId,
-				Move1:           pokemon.Move1,
-				Move2:           pokemon.Move2,
-				Gender:          pokemon.Gender,
-				Cp:              pokemon.Cp,
-				AtkIv:           pokemon.AtkIv,
-				DefIv:           pokemon.DefIv,
-				StaIv:           pokemon.StaIv,
-				//not IvInactive
-				Iv:                      pokemon.Iv,
-				Form:                    pokemon.Form,
-				Level:                   pokemon.Level,
-				EncounterWeather:        pokemon.EncounterWeather, //? perhaps do not include
-				Weather:                 pokemon.Weather,
-				Costume:                 pokemon.Costume,
-				FirstSeenTimestamp:      pokemon.FirstSeenTimestamp,
-				Changed:                 pokemon.Changed,
-				CellId:                  pokemon.CellId,
-				ExpireTimestampVerified: pokemon.ExpireTimestampVerified,
-				DisplayPokemonId:        pokemon.DisplayPokemonId,
-				IsDitto:                 pokemon.IsDitto,
-				SeenType:                pokemon.SeenType,
-				Shiny:                   pokemon.Shiny,
-				Username:                pokemon.Username,
-				Pvp: func() map[string][]gohbem.PokemonEntry {
-					if ohbem != nil {
-						pvp, err := ohbem.QueryPvPRank(int(pokemon.PokemonId),
-							int(pokemon.Form.ValueOrZero()),
-							int(pokemon.Costume.ValueOrZero()),
-							int(pokemon.Gender.ValueOrZero()),
-							int(pokemon.AtkIv.ValueOrZero()),
-							int(pokemon.DefIv.ValueOrZero()),
-							int(pokemon.StaIv.ValueOrZero()),
-							float64(pokemon.Level.ValueOrZero()))
-						if err != nil {
-							return nil
-						}
-						return pvp
-					}
-					return nil
-				}(),
-			}
-
-			results = append(results, &apiPokemon)
-		}
-	}
-
-	log.Infof("GetPokemonInArea - total time %s (locked time %s), %d scanned, %d skipped, %d returned", time.Since(start), lockedTime, pokemonExamined, pokemonSkipped, len(results))
-
-	return results
-}
-
-func GetOnePokemon(pokemonId uint64) *Pokemon {
-	if item := pokemonCache.Get(strconv.FormatUint(pokemonId, 10)); item != nil {
-		pokemon := item.Value()
-		return &pokemon
-	}
-	return nil
-}
-
-func GetAvailablePokemon() []*Available {
-	type pokemonFormKey struct {
-		pokemonId int16
-		form      int16
-	}
-
-	start := time.Now()
-
-	pkmnMap := make(map[pokemonFormKey]int)
-	pokemonLookupCache.Range(func(key uint64, pokemon PokemonLookupCacheItem) bool {
-		pkmnMap[pokemonFormKey{pokemon.PokemonLookup.PokemonId, pokemon.PokemonLookup.Form}]++
-		return true
-	})
-
-	var available []*Available
-	for key, count := range pkmnMap {
-
-		pkmn := &Available{
-			PokemonId: key.pokemonId,
-			Form:      key.form,
-			Count:     count,
-		}
-		available = append(available, pkmn)
-	}
-
-	log.Infof("GetAvailablePokemon - total time %s (locked time --)", time.Since(start))
-
-	return available
-}
-
-func SearchPokemon(request ApiPokemonSearch) []*Pokemon {
-	start := time.Now()
-	results := make([]*Pokemon, 0, request.Limit)
-	pokemonMatched := 0
-
-	if request.SearchIds == nil {
-		return nil
-	}
-
-	pokemonTreeMutex.RLock()
-	pokemonTree2 := pokemonTree.Copy()
-	pokemonTreeMutex.RUnlock()
-
-	maxPokemon := config.Config.Tuning.MaxPokemonResults
-	if request.Limit > 0 && request.Limit < maxPokemon {
-		maxPokemon = request.Limit
-	}
-	pokemonSkipped := 0
-	pokemonScanned := 0
-	maxDistance := float64(1000) // This should come from the request?
-
-	pokemonTree2.Nearby(
-		rtree.BoxDist[float64, uint64]([2]float64{request.Center.Longitude, request.Center.Latitude}, [2]float64{request.Center.Longitude, request.Center.Latitude}, nil),
-		func(min, max [2]float64, pokemonId uint64, dist float64) bool {
-			pokemonLookupItem, inCache := pokemonLookupCache.Load(pokemonId)
-			if !inCache {
-				pokemonSkipped++
-				// Did not find cached result, something amiss?
-				return true
-			}
-
-			pokemonScanned++
-			if dist > maxDistance {
-				log.Infof("SearchPokemon - result would exceed maximum distance (%f), stopping scan", maxDistance)
-				return false
-			}
-
-			found := slices.Contains(request.SearchIds, pokemonLookupItem.PokemonLookup.PokemonId)
-
-			if found {
-				if pokemonCacheEntry := pokemonCache.Get(strconv.FormatUint(pokemonId, 10)); pokemonCacheEntry != nil {
-					pokemon := pokemonCacheEntry.Value()
-					results = append(results, &pokemon)
-					pokemonMatched++
-
-					if pokemonMatched > maxPokemon {
-						log.Infof("SearchPokemon - result would exceed maximum size (%d), stopping scan", maxPokemon)
-						return false
-					}
-				}
-			}
-
-			return true
-		},
-	)
-
-	log.Infof("SearchPokemon - scanned %d pokemon, total time %s, %d returned", pokemonScanned, time.Since(start), len(results))
-	return results
 }
